It now prints!
Some of this is just refactoring. The main difference is proper retrieval of ppd from server, generation of print job, and transfer of print job to cups server.

There is more work to do, such as unlinking the ppd file when done with it. Otherwise releasing resources that are no longer used. Tying the android print job with the cups print job for status updates, etc.