Kill off useless share-to-print feature
If an application can't print, it probably also can't share. Also, even
if it can share, if it can't print, just open the file in an application
that CAN print. This "feature" does nothing but bloat out the share list
menu on applications, which is intended for actually SHARING data, i.e.,
email or other network services.