Cleanups
Delete a bunch of junk we don't need... one can only imagine why the
previous maintainer insisted on having an ABOUT menu from EVERY SINGLE
ACTIVITY... and not to mention, that the about menu didn't actually say
anything besides the application name and version -- neither of which is
relevant information for the end user to have access to while bouncing
around the menus. So that junk is now gone. Also got rid of the
remaining bits of the useless "certificate" menu... which did nothing
that wasn't already handled automatically by CUPS.... so WHY?

Also fixed up the printservice menu so that we have a Settings and an
Add Printer menu item. It was previously weird that the "add printer"
menu opened the settings activity, which in turn, was able to open the
add printer menu. Now we jump directly to the add printer menu, or go to
the settings menu. I think we can actually remove the "add printer" menu
item from the settings activity, since we can now get there directly.